Application Rating Analysis
This notebook conducts an analysis of applicants ratings. It includes data cleaning steps such as removing duplicate applicant IDs, filling missing values in the 'External Rating' and 'Education level' fields, and merging data. The application rating is calculated based on specific criteria, resulting in a score ranging from 0 to 100. The notebook filters out applications with a rating of zero and groups the remaining accepted applications by the week of submission. It then visualizes the average rating of accepted applications for each week on a graph.